Support Plugin handling

+++ Assets/Plugins/Editor/Uplift/Schemas/Extensions/InstalledPackage.cs
@@ -2,7 +2,9 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using UnityEditor;
 using UnityEngine;
+using Uplift.Common;
 
 namespace Uplift.Schemas
 {
@@ -12,26 +14,33 @@ namespace Uplift.Schemas
         {
             var dirPaths = new List<string>();
             
-            foreach (InstallationSpecs spec in Install)
+            foreach (InstallSpec spec in Install)
             {
-                if (spec.Kind == InstallSpecType.Root)
+                if (spec.Type == InstallSpecType.Root)
                 {
                     // Removing Root package
                     Directory.Delete(spec.Path, true);
                 }
                 else
                 {
-                    var sourceDir = UpfileHandler.Instance().GetDestinationFor(spec.Kind).Location;
+                    var sourceDir = UpfileHandler.Instance().GetDestinationFor(spec).Location;
 
-                    var filePath = Path.Combine(sourceDir, spec.Path);
+                    //var filePath = Path.Combine(sourceDir, spec.Path);
+                    // Due to change, filePath right now IS spec.Path
+                    var filePath = spec.Path;
 
                     try
                     {
                         File.Delete(filePath);
+                        File.Delete(filePath + ".meta"); // Removing meta files as well.
                     }
                     catch (FileNotFoundException)
                     {
                         Debug.Log("Warning, tracked file not found: " + filePath);
+                    }
+                    catch (DirectoryNotFoundException)
+                    {
+                        Debug.Log("Warning, tracked directory not found: " + filePath);
                     }
 
 
@@ -40,7 +49,7 @@ namespace Uplift.Schemas
                     if (!string.IsNullOrEmpty(dirName))
                     {
 
-                        dirPaths.Add(Path.Combine(sourceDir, dirName));
+                        dirPaths.Add(dirName);
                     }
                     
                     
@@ -49,22 +58,61 @@ namespace Uplift.Schemas
                 
             }
 
-            foreach (var p in dirPaths.Distinct())
+            // An itchy bit.
+            // Paths can nest. Ordering of paths is not guaranteed.
+            // So we'll loop over removing dirs to the point, where no action has been made.
+            
+            var actionsDone = 1;
+            var loopCounter = 1;
+
+            while (actionsDone != 0)
             {
-                if(string.IsNullOrEmpty(p))
+                if (loopCounter > 5)
                 {
-                    continue;
+                    Debug.LogWarning(
+                        "Warning: Nuke Dependency Loop has done more than 5 rounds. This might or might not be error, depending on setup"
+                        );
+                    
+                    
                 }
                 
-                try
+                actionsDone = 0;
+                loopCounter++;
+
+                // We're recursively listing the directories we're keeping so that we can extract empty directories.
+                var recursiveDirPaths = FileSystemUtil.RecursivelyDirPaths(dirPaths).Distinct().ToList();
+
+                foreach (var p in recursiveDirPaths)
                 {
-                    Debug.Log("Removing " + p);
-                    Directory.Delete(p, true);
-                }
-                catch (DirectoryNotFoundException)
-                {
-                    // Few places where this might happen:
-                    // First place
+                    
+                    if (string.IsNullOrEmpty(p))
+                    {
+                        continue;
+                    }
+
+                    if (!Directory.Exists(p))
+                    {
+                        continue; 
+                    }
+                    
+                    var filesInDirectory = Directory.GetFiles(p).Length;
+                    var directoriesInDirectory = Directory.GetDirectories(p).Length;
+
+                    if (filesInDirectory + directoriesInDirectory > 0)
+                    {
+                        continue;
+                    }
+
+                    try
+                    {
+                        Directory.Delete(p, true);
+                        actionsDone += 1;
+                        File.Delete(p + ".meta"); // .meta file for directory might exist, remove it as well
+                    }
+                    catch (DirectoryNotFoundException)
+                    {
+                        
+                    }
                 }
             }
         }
